+++ .env
@@ -1,0 +1,1 @@
+ESLINT_NO_DEV_ERRORS=true
